Move *.test.js
