+++ src/lib/reversedUnixtime.test.js
@@ -2,7 +2,7 @@
 const describe = require('mocha').describe;
 const it = require('mocha').it;
 const assert = require('power-assert');
-const reversedUnixtime = require('../../src/lib/reversedUnixtime');
+const { reversedUnixtime } = require('./functions');
 
 describe('reversedUnixtime', () => {
     it ('reversedUnixtime(0) -> "9007199254740991"', () => {
